+++ packages/prisma/src/migrations/20241228211802_order_add_stop_price/migration.sql
@@ -1,2 +1,3 @@
 -- AlterTable
-ALTER TABLE "Order" ADD COLUMN "stopPrice" REAL;
+ALTER TABLE "Order"
+  ADD COLUMN "stopPrice" DOUBLE PRECISION;
